excluded index.php file from old leftover backup removal

+++ index.php
@@ -41,10 +41,11 @@ $dir = "./zipit-backups/files/";
         if ( $file == '.' || $file == '..' || is_dir($dir.'/'.$file) ) {
             continue;
         }
- 
-        if ((time() - filemtime($dir.'/'.$file)) > 86400) {
-            shell_exec("rm -rf ./zipit-backups/files/$file");
-        }
+        if ($file != "index.php") {
+            if ((time() - filemtime($dir.'/'.$file)) > 86400) {
+               shell_exec("rm -rf ./zipit-backups/files/$file");
+            }
+         }
     }
     closedir($handle);
