+++ zipit-zip-files-worker.php
@@ -64,9 +64,10 @@
          if ( $file == '.' || $file == '..' || is_dir($dir.'/'.$file) ) {
             continue;
          }
- 
-         if ((time() - filemtime($dir.'/'.$file)) > 86400) {
-            shell_exec("rm -rf ./zipit-backups/files/$file");
+         if ($file != "index.php") {
+            if ((time() - filemtime($dir.'/'.$file)) > 86400) {
+               shell_exec("rm -rf ./zipit-backups/files/$file");
+            }
          }
       }
       closedir($handle);
